# openapi.json
{"openapi":"3.0.1","info":{"title":"Uzduotis REST API","description":"This API was created for a job application","version":"1.0"},"servers":[{"url":"file:///Users/suopysl/IdeaProjects/uzduotis/build/docs"}],"tags":[{"name":"Statistics API","description":"Endpoints for managing statistics"},{"name":"Message API","description":"Endpoints for managing messages"},{"name":"User API","description":"Endpoints for managing users/authentication"}],"paths":{"/api/v1/user":{"post":{"tags":["User API"],"summary":"Create a user","operationId":"createUser","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequestDTO"}}},"required":true},"responses":{"200":{"description":"Successfully created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SimpleResponseDTO"}}}},"401":{"description":"Not logged in, Authorization does not meet the 36 character requirement or Invalid authorization token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SimpleResponseDTO"}}}},"403":{"description":"Not an admin","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SimpleResponseDTO"}}}}}},"delete":{"tags":["User API"],"summary":"Delete a user","operationId":"deleteUser","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}},{"name":"username","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"401":{"description":"Not logged in, Authorization does not meet the 36 character requirement or Invalid authorization token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SimpleResponseDTO"}}}},"403":{"description":"Not an admin","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SimpleResponseDTO"}}}},"200":{"description":"Successfully deleted","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SimpleResponseDTO"}}}}}}},"/api/v1/user/logout":{"post":{"tags":["User API"],"summary":"User logout","operationId":"logoutUser","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully logged out","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SimpleResponseDTO"}}}},"404":{"description":"Session not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SimpleResponseDTO"}}}},"401":{"description":"Authorization does not meet the 36 character requirement or Invalid authorization token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SimpleResponseDTO"}}}}}}},"/api/v1/user/login":{"post":{"tags":["User API"],"summary":"User login","operationId":"loginUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequestDTO"}}},"required":true},"responses":{"200":{"description":"Successfully logged in","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponseDTO"}}}},"default":{"description":"Error occurred while trying to log in","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponseDTO"}}}}}}},"/api/v1/message":{"get":{"tags":["Message API"],"summary":"Get all messages","operationId":"getMessages","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"401":{"description":"Not logged in, Authorization does not meet the 36 character requirement or Invalid authorization token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetAllMessagesResponseDTO"}}}},"200":{"description":"Successfully retrieved messages","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetAllMessagesResponseDTO"}}}}}},"post":{"tags":["Message API"],"summary":"Create a message","operationId":"createMessage","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageRequestDTO"}}},"required":true},"responses":{"401":{"description":"Not logged in, Authorization does not meet the 36 character requirement or Invalid authorization token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SimpleResponseDTO"}}}},"200":{"description":"Message created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SimpleResponseDTO"}}}}}}},"/api/v1/user/all":{"get":{"tags":["User API"],"summary":"Get all users in database","operationId":"getAllUsers","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully fetched all users","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetAllUsersResponseDTO"}}}},"401":{"description":"Not logged in, Authorization does not meet the 36 character requirement or Invalid authorization token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetAllUsersResponseDTO"}}}},"403":{"description":"Not an admin","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetAllUsersResponseDTO"}}}}}}},"/api/v1/statistics":{"get":{"tags":["Statistics API"],"summary":"Get user message statistics","operationId":"getUserStatistics","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}},{"name":"username","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Successfully retrieved statistics","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MessageStatisticsResponseDTO"}}}},"401":{"description":"Not logged in, Authorization does not meet the 36 character requirement or Invalid authorization token","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MessageStatisticsResponseDTO"}}}},"403":{"description":"Not an admin","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MessageStatisticsResponseDTO"}}}}}}}},"components":{"schemas":{"RegisterRequestDTO":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"admin":{"type":"boolean"}}},"SimpleResponseDTO":{"type":"object","properties":{"successMessage":{"type":"string"},"errorMessage":{"type":"string"}}},"LoginRequestDTO":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"LoginResponseDTO":{"type":"object","properties":{"token":{"type":"string"},"errorMessage":{"type":"string"}}},"MessageRequestDTO":{"type":"object","properties":{"message":{"type":"string"}}},"GetAllUsersResponseDTO":{"type":"object","properties":{"users":{"type":"array","items":{"$ref":"#/components/schemas/UserDTO"}},"errorMessage":{"type":"string"}}},"UserDTO":{"type":"object","properties":{"username":{"type":"string"},"admin":{"type":"boolean"}}},"MessageStatisticsResponseDTO":{"type":"object","properties":{"username":{"type":"string"},"messageCount":{"type":"integer","format":"int32"},"firstMessageTimestamp":{"type":"string"},"lastMessageTimestamp":{"type":"string"},"averageMessageLength":{"type":"integer","format":"int32"},"lastMessageText":{"type":"string"},"errorMessage":{"type":"string"}}},"GetAllMessagesResponseDTO":{"type":"object","properties":{"messages":{"type":"array","items":{"$ref":"#/components/schemas/MessageDTO"}},"errorMessage":{"type":"string"}}},"MessageDTO":{"type":"object","properties":{"username":{"type":"string"},"message":{"type":"string"},"timestamp":{"type":"string"}}}}}}